pyproject.toml
[project]
name = "tiktok-pipeline-extension"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.0",
    "langgraph>=1.1.9",
    "moviepy>=2.2.1",
    "opencv-python>=4.13.0.92",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "uvicorn>=0.46.0",
    "yt-dlp>=2026.3.17",
]
